Moved aptslav dto's in corresponding package
