+++ api/src/main/java/com/eleks/academy/pharmagator/dataproviders/AptslavDataProvider.java
@@ -1,9 +1,9 @@
 package com.eleks.academy.pharmagator.dataproviders;
 
-import com.eleks.academy.pharmagator.dataproviders.aptslav.dto.AptslavMedicineDto;
-import com.eleks.academy.pharmagator.dataproviders.aptslav.dto.AptslavResponseBody;
-import com.eleks.academy.pharmagator.dataproviders.aptslav.dto.converters.ApiDtoConverter;
 import com.eleks.academy.pharmagator.dataproviders.dto.MedicineDto;
+import com.eleks.academy.pharmagator.dataproviders.dto.aptslav.AptslavMedicineDto;
+import com.eleks.academy.pharmagator.dataproviders.dto.aptslav.AptslavResponseBody;
+import com.eleks.academy.pharmagator.dataproviders.dto.aptslav.converters.ApiDtoConverter;
 import com.eleks.academy.pharmagator.repositories.PharmacyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -104,4 +104,5 @@ public class AptslavDataProvider implements DataProvider {
                 })
                 .block();
     }
+
 }
