+++ api/src/main/java/com/eleks/academy/pharmagator/dataproviders/dto/aptslav/AptslavPriceDto.java
@@ -1,0 +1,14 @@
+package com.eleks.academy.pharmagator.dataproviders.dto.aptslav;
+
+import lombok.Data;
+
+import java.math.BigDecimal;
+
+@Data
+public class AptslavPriceDto {
+
+    private BigDecimal max;
+
+    private BigDecimal min;
+
+}
