+++ api/src/main/java/com/eleks/academy/pharmagator/dataproviders/dto/aptslav/converters/ApiDtoConverter.java
@@ -1,4 +1,4 @@
-package com.eleks.academy.pharmagator.dataproviders.aptslav.dto.converters;
+package com.eleks.academy.pharmagator.dataproviders.dto.aptslav.converters;
 
 import com.eleks.academy.pharmagator.dataproviders.dto.MedicineDto;
