+++ api/src/main/java/com/eleks/academy/pharmagator/dataproviders/dto/ds/DSMedicineDto.java
@@ -1,18 +1,19 @@
 package com.eleks.academy.pharmagator.dataproviders.dto.ds;
 
-import com.eleks.academy.pharmagator.dataproviders.dto.ds.dto.DSMedicineDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class MedicinesResponse {
+public class DSMedicineDto {
 
-	private Long total;
-	private List<DSMedicineDto> products;
+    private String id;
+    private String name;
+    private BigDecimal price;
+    private String manufacturer;
 
 }
