--- api/src/main/java/com/eleks/academy/pharmagator/dataproviders/dto/ds/dto/DSMedicineDto.java
@@ -1,19 +1,0 @@
-package com.eleks.academy.pharmagator.dataproviders.dto.ds.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class DSMedicineDto {
-
-    private String id;
-    private String name;
-    private BigDecimal price;
-    private String manufacturer;
-
-}
